% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2021_03_22.R
\docType{data}
\name{rcd_data_2021_03_22}
\alias{rcd_data_2021_03_22}
\alias{rcd_2021_03_22_adsl}
\alias{rcd_2021_03_22_adae}
\alias{rcd_2021_03_22_adaette}
\alias{rcd_2021_03_22_adcm}
\alias{rcd_2021_03_22_addv}
\alias{rcd_2021_03_22_adeg}
\alias{rcd_2021_03_22_adex}
\alias{rcd_2021_03_22_adlb}
\alias{rcd_2021_03_22_admh}
\alias{rcd_2021_03_22_adqs}
\alias{rcd_2021_03_22_adrs}
\alias{rcd_2021_03_22_adsub}
\alias{rcd_2021_03_22_adtr}
\alias{rcd_2021_03_22_adtte}
\alias{rcd_2021_03_22_advs}
\title{Cached datasets from the 2021_03_22 release of \code{random.cdisc.data}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 38 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 65 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2800 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3602 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 71 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 77 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 48 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 56 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 48 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 48 columns.

An object of class \code{data.frame} with 2800 rows and 60 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1600 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 68 columns.
}
\usage{
rcd_2021_03_22_adsl

rcd_2021_03_22_adae

rcd_2021_03_22_adaette

rcd_2021_03_22_adcm

rcd_2021_03_22_addv

rcd_2021_03_22_adeg

rcd_2021_03_22_adex

rcd_2021_03_22_adlb

rcd_2021_03_22_admh

rcd_2021_03_22_adqs

rcd_2021_03_22_adrs

rcd_2021_03_22_adsub

rcd_2021_03_22_adtr

rcd_2021_03_22_adtte

rcd_2021_03_22_advs
}
\description{
Cached datasets from the 2021_03_22 release of \code{random.cdisc.data}
}
\keyword{datasets}
