% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2021_10_13.R
\docType{data}
\name{rcd_data_2021_10_13}
\alias{rcd_data_2021_10_13}
\alias{rcd_2021_10_13_adsl}
\alias{rcd_2021_10_13_adae}
\alias{rcd_2021_10_13_adaette}
\alias{rcd_2021_10_13_adcm}
\alias{rcd_2021_10_13_addv}
\alias{rcd_2021_10_13_adeg}
\alias{rcd_2021_10_13_adex}
\alias{rcd_2021_10_13_adhy}
\alias{rcd_2021_10_13_adlb}
\alias{rcd_2021_10_13_admh}
\alias{rcd_2021_10_13_adqs}
\alias{rcd_2021_10_13_adrs}
\alias{rcd_2021_10_13_adsub}
\alias{rcd_2021_10_13_adtr}
\alias{rcd_2021_10_13_adtte}
\alias{rcd_2021_10_13_advs}
\title{Cached datasets from the 2021_10_13 release of \code{random.cdisc.data}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 39 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 69 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3600 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3602 rows and 57 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13600 rows and 72 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 51 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 55 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 78 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14000 rows and 57 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 49 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 49 columns.

An object of class \code{data.frame} with 2800 rows and 61 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1600 rows and 50 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 69 columns.
}
\usage{
rcd_2021_10_13_adsl

rcd_2021_10_13_adae

rcd_2021_10_13_adaette

rcd_2021_10_13_adcm

rcd_2021_10_13_addv

rcd_2021_10_13_adeg

rcd_2021_10_13_adex

rcd_2021_10_13_adhy

rcd_2021_10_13_adlb

rcd_2021_10_13_admh

rcd_2021_10_13_adqs

rcd_2021_10_13_adrs

rcd_2021_10_13_adsub

rcd_2021_10_13_adtr

rcd_2021_10_13_adtte

rcd_2021_10_13_advs
}
\description{
Cached datasets from the 2021_10_13 release of \code{random.cdisc.data}
}
\keyword{datasets}
